Added : Changes

+++ ModelsLibrary/Models/ResponseModel/Response.cs
@@ -3,7 +3,7 @@
     public class Response<T>
     {
         public string Message { get; set; }
-        public int Status { get; set; }
+        public long Status { get; set; }
         public T Data { get; set; }
     }
 }
